+++ sentry-types/src/auth.rs
@@ -1,5 +1,3 @@
-#[allow(unused_imports, deprecated)]
-use std::ascii::AsciiExt;
 use std::borrow::Cow;
 use std::fmt;
 use std::str::FromStr;